requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-redis==5.4.0
django-soft-delete==0.9.21
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
isodate==0.6.1
lxml==5.1.0
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
persiantools==3.0.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
zeep==4.2.1
